refactor(cli): extract OAuth toml patchers to shared module
Move addProviderToOAuthTier, removeProviderFromOAuthTier, setProviderAsPrimary,
and the readOAuthTier helper out of login.ts into a new src/cli/utils/toml-patchers.ts.
login.ts re-exports them so existing tests and callers continue to work until
the upcoming command rename and login.ts deletion.
